+++ Tweakr.php
@@ -2,8 +2,8 @@
 /**
     Plugin Name: Tweakr Toolkit
     Plugin URI: https://andidittrich.de/go/tweakr
-    Description: Extends your WordPress Blog with a bunch of common Tweaks and Utilities
-    Version: 1.2-BETA1
+    Description: Supercharges your Blog with common used Tweaks, Features and Utilities
+    Version: 1.2
     Author: Andi Dittrich
     Author URI: https://andidittrich.de
     License: MIT X11 License
@@ -23,7 +23,7 @@
     Plugin Bootstrap Operation
 */
 define('TWEAKR_INIT', true);
-define('TWEAKR_VERSION', '1.2-BETA1');
+define('TWEAKR_VERSION', '1.2');
 define('TWEAKR_PLUGIN_PATH', dirname(__FILE__));
 define('TWEAKR_PLUGIN_URL', plugins_url('/tweakr/'));