+++ class/Tweakr.php
@@ -319,14 +319,14 @@ class Tweakr{
             // plugin installed ?
             if ($version == '0.0.0'){
                 // store new version
-                update_option('tweakr-version', '1.2-BETA1');
+                update_option('tweakr-version', '1.2');
 
             // plugin upgraded ?
-            }else if (version_compare('1.2-BETA1', $version, '>')){
+            }else if (version_compare('1.2', $version, '>')){
                 // run upgrade hook
                 if ($i->_wp_plugin_upgrade($version)){
                     // store new version
-                    update_option('tweakr-version', '1.2-BETA1');
+                    update_option('tweakr-version', '1.2');
 
                     // set flag (string!)
                     update_option('tweakr-upgrade', 'true');