+++ CHANGES.md
@@ -3,9 +3,13 @@
 ### 1.2 ###
 * Added: VisualComponents Extension to visualize the HTML Element Structure in Visual Editor Mode (headings,section,p)
 * Added: Option to remove Rewrite Rules for all types of Feeds (RSS, RTF, ATOM, RSS2)
+* Added: Option to add `.html` extensions to **Pages** or **Categories**
+* Changed: Settings Page Structure has been modified
+* Changed: Updated Piwik and Google Analytics Code
 * Changed: "Disable XMLRPC" will now force the `xmlrpc.php` Endpoint to return an HTTP403 Response
 * Changed: TinyMCE Autowidth is set to **95%**
 * Changed: **Disable oEmbeds** removes also the related Rewrite Rules as well as the endpoint (embed.php template page)
+* Changed: Renamed the setting of "Fix Mail-From" - has to be activated again
 
 ### 1.1 ###
 * Added: Google Analytics Tracking Option