+++ CHANGES.md
@@ -7,8 +7,8 @@
 * Added: Option to add `.html` extensions to **Pages** or **Categories**
 * Added: E-Mail settings to set the mail-from-address as well as mail-from-name manually
 * Changed: Settings Page Structure has been modified
-* Changed: Updated Piwik and Google Analytics Code
-* Changed: "Disable XMLRPC" will now force the `xmlrpc.php` Endpoint to return an HTTP403 Response
+* Changed: New Piwik and Google Analytics Code
+* Changed: "Disable XMLRPC" will now force the `xmlrpc.php` Endpoint to return a HTTP403 Response
 * Changed: TinyMCE Autowidth is set to **95%**
 * Changed: **Disable oEmbeds** removes also the related Rewrite Rules as well as the endpoint (embed.php template page)
 * Changed: Renamed the setting of "Fix Mail-From" - has to be activated again